requirements.txt
celery[redis]>=3.1.19
Flask>=0.10.1
Flask-SQLAlchemy>=2.1
google-api-python-client==1.4.2
oauth2client==1.5.2
gunicorn>=19.1
numba>=0.22.1
numpy>=1.10.1
pandas>=0.17.0
Pillow>=3.0.0
psycopg2>=2.6.1
pycrypto>=2.6.1
pyOpenSSL>=0.15.1
requests>=2.8.1
responses
SQLAlchemy>=1.0.9
